Crate netsim_embed

Modules

arp

ARP packet abstraction.

ethernet

An ethernet packet abstraction.

gre

Minimal GRE Packet implementation: suitable for inspection not generation (e.g. checksum not implemented).

icmp

An ICMP packet abstraction.

icmpv6

An ICMPv6 packet abstraction.

ip

Defines the type and constants for IP next header/next level protocol fields.

ipv4

An IPv4 packet abstraction.

ipv6

An IPv6 packet abstraction.

quic

tcp

A TCP packet abstraction.

types

Provides type aliases for various primitive integer types

udp

A UDP packet abstraction.

util

Utilities for working with packets, eg. checksumming.

vlan

A VLAN packet abstraction.

Structs

DelayBuffer

Ipv4Range

A range of IPv4 addresses with a common prefix

MacAddr

A MAC address.

Machine

Spawns a thread in a new network namespace and configures a TUN interface that sends and receives IP packets from the tx/rx channels and runs some UDP/TCP networking code in task.

MachineId

Namespace

NatConfig

Netsim

Network

NetworkId

Enums

MutPacketData

Mutable packet data.

PacketData

Packet data.

Protocol

Constants

ETHER_ADDR_LEN

The number of bytes in an ethernet (MAC) address.

Traits

FromPacket

Used to convert on-the-wire packets to their #[packet] equivalent.

MutablePacket

Represents a generic, mutable, network packet.

Packet

Represents a generic network packet.

PacketSize

Used to find the calculated size of the packet. This is used for occasions where the underlying buffer is not the same length as the packet itself.

PrimitiveValues

Used to convert a type to primitive values representing it.

Functions

run

unshare_user

Derive Macros

Packet

The entry point for the derive(Packet) custom derive