Module netsim_embed::quic

Modules

FrameTypes

HeaderForms

LongPacketTypes

Structs

Ack

AckIterable

Used to iterate over a slice of {name}Packets

AckPacket

A structure enabling manipulation of on the wire packets

ConnectionClose

ConnectionCloseIterable

Used to iterate over a slice of {name}Packets

ConnectionClosePacket

A structure enabling manipulation of on the wire packets

Crypto

CryptoIterable

Used to iterate over a slice of {name}Packets

CryptoPacket

A structure enabling manipulation of on the wire packets

DataBlocked

DataBlockedIterable

Used to iterate over a slice of {name}Packets

DataBlockedPacket

A structure enabling manipulation of on the wire packets

FixedBit

FrameType

Handshake

HandshakeDone

HandshakeDoneIterable

Used to iterate over a slice of {name}Packets

HandshakeDonePacket

A structure enabling manipulation of on the wire packets

HandshakeIterable

Used to iterate over a slice of {name}Packets

HandshakePacket

A structure enabling manipulation of on the wire packets

HeaderForm

Initial

InitialIterable

Used to iterate over a slice of {name}Packets

InitialPacket

A structure enabling manipulation of on the wire packets

LongPacketType

MaxData

MaxDataIterable

Used to iterate over a slice of {name}Packets

MaxDataPacket

A structure enabling manipulation of on the wire packets

MaxStreamData

MaxStreamDataIterable

Used to iterate over a slice of {name}Packets

MaxStreamDataPacket

A structure enabling manipulation of on the wire packets

MaxStreams

MaxStreamsIterable

Used to iterate over a slice of {name}Packets

MaxStreamsPacket

A structure enabling manipulation of on the wire packets

MutableAckPacket

A structure enabling manipulation of on the wire packets

MutableConnectionClosePacket

A structure enabling manipulation of on the wire packets

MutableCryptoPacket

A structure enabling manipulation of on the wire packets

MutableDataBlockedPacket

A structure enabling manipulation of on the wire packets

MutableHandshakeDonePacket

A structure enabling manipulation of on the wire packets

MutableHandshakePacket

A structure enabling manipulation of on the wire packets

MutableInitialPacket

A structure enabling manipulation of on the wire packets

MutableMaxDataPacket

A structure enabling manipulation of on the wire packets

MutableMaxStreamDataPacket

A structure enabling manipulation of on the wire packets

MutableMaxStreamsPacket

A structure enabling manipulation of on the wire packets

MutableNewConnectionIdPacket

A structure enabling manipulation of on the wire packets

MutableNewTokenPacket

A structure enabling manipulation of on the wire packets

MutableOneRttPacket

A structure enabling manipulation of on the wire packets

MutablePaddingPacket

A structure enabling manipulation of on the wire packets

MutablePathChallengePacket

A structure enabling manipulation of on the wire packets

MutablePathResponsePacket

A structure enabling manipulation of on the wire packets

MutablePingPacket

A structure enabling manipulation of on the wire packets

MutableResetStreamPacket

A structure enabling manipulation of on the wire packets

MutableRetireConnectionIdPacket

A structure enabling manipulation of on the wire packets

MutableRetryPacket

A structure enabling manipulation of on the wire packets

MutableStopSendingPacket

A structure enabling manipulation of on the wire packets

MutableStreamDataBlockedPacket

A structure enabling manipulation of on the wire packets

MutableStreamPacket

A structure enabling manipulation of on the wire packets

MutableStreamsBlockedPacket

A structure enabling manipulation of on the wire packets

MutableVersionNegotiationPacket

A structure enabling manipulation of on the wire packets

MutableZeroRttPacket

A structure enabling manipulation of on the wire packets

NewConnectionId

NewConnectionIdIterable

Used to iterate over a slice of {name}Packets

NewConnectionIdPacket

A structure enabling manipulation of on the wire packets

NewToken

NewTokenIterable

Used to iterate over a slice of {name}Packets

NewTokenPacket

A structure enabling manipulation of on the wire packets

OneRtt

OneRttIterable

Used to iterate over a slice of {name}Packets

OneRttPacket

A structure enabling manipulation of on the wire packets

Padding

PaddingIterable

Used to iterate over a slice of {name}Packets

PaddingPacket

A structure enabling manipulation of on the wire packets

PathChallenge

PathChallengeIterable

Used to iterate over a slice of {name}Packets

PathChallengePacket

A structure enabling manipulation of on the wire packets

PathResponse

PathResponseIterable

Used to iterate over a slice of {name}Packets

PathResponsePacket

A structure enabling manipulation of on the wire packets

Ping

PingIterable

Used to iterate over a slice of {name}Packets

PingPacket

A structure enabling manipulation of on the wire packets

ResetStream

ResetStreamIterable

Used to iterate over a slice of {name}Packets

ResetStreamPacket

A structure enabling manipulation of on the wire packets

RetireConnectionId

RetireConnectionIdIterable

Used to iterate over a slice of {name}Packets

RetireConnectionIdPacket

A structure enabling manipulation of on the wire packets

Retry

RetryIterable

Used to iterate over a slice of {name}Packets

RetryPacket

A structure enabling manipulation of on the wire packets

StopSending

StopSendingIterable

Used to iterate over a slice of {name}Packets

StopSendingPacket

A structure enabling manipulation of on the wire packets

Stream

StreamDataBlocked

StreamDataBlockedIterable

Used to iterate over a slice of {name}Packets

StreamDataBlockedPacket

A structure enabling manipulation of on the wire packets

StreamIterable

Used to iterate over a slice of {name}Packets

StreamPacket

A structure enabling manipulation of on the wire packets

StreamsBlocked

StreamsBlockedIterable

Used to iterate over a slice of {name}Packets

StreamsBlockedPacket

A structure enabling manipulation of on the wire packets

Version

VersionNegotiation

VersionNegotiationIterable

Used to iterate over a slice of {name}Packets

VersionNegotiationPacket

A structure enabling manipulation of on the wire packets

ZeroRtt

ZeroRttIterable

Used to iterate over a slice of {name}Packets

ZeroRttPacket

A structure enabling manipulation of on the wire packets

Enums

Frame

QuicPacket

Functions

n_varints

packet_number

varint

varint_length